moving bf1 to scripts directory. function in fit.py
